Refactor views to screens
