+++ lib/authentication/authentication.dart
@@ -2,26 +2,3 @@ export 'bloc/authentication_bloc.dart';
 export 'enums/enums.dart';
 export 'login/login.dart';
 export 'models/models.dart';
-
-// export 'sign_up/sign_up.dart';
-
-// import 'package:firebase_auth/firebase_auth.dart';
-// import 'package:shopping_list/globals.dart';
-
-// void initAuth() {
-//   Globals.auth = FirebaseAuth.instance;
-
-//   subscribeAuth();
-// }
-
-// void subscribeAuth() {
-//   Globals.auth.authStateChanges().listen((User? user) {
-//     if (user == null) {
-//       print('User is currently signed out!');
-//       Globals.user = null;
-//     } else {
-//       print('User is signed in!');
-//       Globals.user = user;
-//     }
-//   });
-// }
