--- lib/authentication/login/view/login_screen.dart
@@ -1,23 +1,0 @@
-import 'package:authentication_repository/authentication_repository.dart';
-import 'package:flutter/material.dart';
-import 'package:flutter_bloc/flutter_bloc.dart';
-
-import '../../authentication.dart';
-
-class LoginScreen extends StatelessWidget {
-  static const id = 'login_screen';
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(title: const Text('Login')),
-      body: Padding(
-        padding: const EdgeInsets.all(8.0),
-        child: BlocProvider(
-          create: (_) => LoginCubit(context.read<AuthenticationRepository>()),
-          child: LoginForm(),
-        ),
-      ),
-    );
-  }
-}
