+++ lib/home/home.dart
@@ -1,2 +1,2 @@
-export 'view/home_screen.dart';
+export 'screens/screens.dart';
 export 'widgets/widgets.dart';
