--- lib/splash/splash.dart
@@ -1,1 +1,0 @@
-export 'view/splash_screen.dart';
